Added serial interface.
